restructure-component-user
