+++ src/components/pages/UserPage/components/Aurora.tsx
