+++ src/components/pages/UserPage/components/Friend.tsx
