+++ src/components/pages/UserPage/index.tsx
@@ -1,6 +1,6 @@
-import Achievement from "@/components/User/achievement";
-import UserTitle from "@/components/User/UserTitle";
-import Friend from "@/components/User/Friend";
+import Achievement from "@/components/pages/UserPage/components/achievement";
+import UserTitle from "@/components/pages/UserPage/components/UserTitle";
+import Friend from "@/components/pages/UserPage/components/Friend";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { usePostLogOutSupa, useGetUserSupa } from "@/api";
 import { useEffect } from "react";
